update: rename default header file, create detail header file and update imports
