+++ src/app/components/Headers/HeaderRecipeDetail.tsx
@@ -1,0 +1,6 @@
+import { AppBar } from '@mui/material';
+import React from 'react';
+
+export default (): JSX.Element => {
+  return <AppBar></AppBar>;
+};
