+++ src/app/components/Headers/header.stories.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Header from '.';
+import Header from './HeaderDefault';
 
 export default {
   title: 'Component/Header',
