+++ src/app/pages/Inventory/Inventory.tsx
@@ -4,7 +4,7 @@ import { CssBaseline, List } from '@mui/material';
 import React, { useState } from 'react';
 import { useSelector } from 'react-redux';
 import Footer from '../../components/Footer';
-import Header from '../../components/Header';
+import Header from '../../components/Headers/HeaderDefault';
 import InventoryItem from '../../components/InventoryItem';
 import ListForm from '../../components/ListForm';
 import { addInventoryItem } from '../../reducers/inventoryReducer';
